+++ src/tests/test.index.js
@@ -1,7 +1,7 @@
 const IpcProvider = require('../index.js'); // eslint-disable-line
-const TestRPC = require('ethereumjs-testrpc'); // eslint-disable-line
-const Eth = require('ethjs-query'); // eslint-disable-line
-const EthQuery = require('eth-query');
+const TestRPC = require('vaporyjs-testrpc'); // eslint-disable-line
+const Vap = require('vapjs-query'); // eslint-disable-line
+const VapQuery = require('vap-query');
 const Web3 = require('web3');
 const assert = require('chai').assert; // eslint-disable-line
 const SandboxedModule = require('sandboxed-module');
@@ -103,11 +103,11 @@ describe('IpcProvider', () => {
     });
   });
 
-  describe('test against ethjs-query', () => {
-    const eth = new Eth(new IpcProvider('/tmp/out.ipc')); // eslint-disable-line
+  describe('test against vapjs-query', () => {
+    const vap = new Vap(new IpcProvider('/tmp/out.ipc')); // eslint-disable-line
 
     it('should get accounts', (done) => {
-      eth.accounts((accountsError, accountsResult) => {
+      vap.accounts((accountsError, accountsResult) => {
         assert.equal(accountsError, null);
         assert.equal(typeof accountsResult, 'object');
         assert.equal(Array.isArray(accountsResult), true);
@@ -117,12 +117,12 @@ describe('IpcProvider', () => {
     });
 
     it('should get balances', (done) => {
-      eth.accounts((accountsError, accountsResult) => {
+      vap.accounts((accountsError, accountsResult) => {
         assert.equal(accountsError, null);
         assert.equal(typeof accountsResult, 'object');
         assert.equal(Array.isArray(accountsResult), true);
 
-        eth.getBalance(accountsResult[0], (balanceError, balanceResult) => {
+        vap.getBalance(accountsResult[0], (balanceError, balanceResult) => {
           assert.equal(balanceError, null);
           assert.equal(typeof balanceResult, 'object');
           assert.equal(balanceResult.toNumber(10) > 0, true);
@@ -133,11 +133,11 @@ describe('IpcProvider', () => {
     });
 
     it('should get coinbase and balance', (done) => {
-      eth.coinbase((accountsError, accountResult) => {
+      vap.coinbase((accountsError, accountResult) => {
         assert.equal(accountsError, null);
         assert.equal(typeof accountResult, 'string');
 
-        eth.getBalance(accountResult, (balanceError, balanceResult) => {
+        vap.getBalance(accountResult, (balanceError, balanceResult) => {
           assert.equal(balanceError, null);
           assert.equal(typeof balanceResult, 'object');
           assert.equal(balanceResult.toNumber(10) > 0, true);
@@ -148,8 +148,8 @@ describe('IpcProvider', () => {
     });
   });
 
-  describe('test against eth-query', () => {
-    const query = new EthQuery(new IpcProvider('/tmp/out.ipc')); // eslint-disable-line
+  describe('test against vap-query', () => {
+    const query = new VapQuery(new IpcProvider('/tmp/out.ipc')); // eslint-disable-line
 
     it('should get accounts', (done) => {
       query.accounts((accountsError, accountsResult) => {
@@ -195,7 +195,7 @@ describe('IpcProvider', () => {
     const web3 = new Web3(new IpcProvider('/tmp/out.ipc')); // eslint-disable-line
 
     it('should get accounts WEB3', (done) => {
-      web3.eth.getAccounts((accountsError, accountsResult) => {
+      web3.vap.getAccounts((accountsError, accountsResult) => {
         assert.equal(accountsError, null);
         assert.equal(typeof accountsResult, 'object');
         assert.equal(Array.isArray(accountsResult), true);
@@ -204,12 +204,12 @@ describe('IpcProvider', () => {
     });
 
     it('should get balances', (done) => {
-      web3.eth.getAccounts((accountsError, accountsResult) => {
+      web3.vap.getAccounts((accountsError, accountsResult) => {
         assert.equal(accountsError, null);
         assert.equal(typeof accountsResult, 'object');
         assert.equal(Array.isArray(accountsResult), true);
 
-        web3.eth.getBalance(accountsResult[0], (balanceError, balanceResult) => {
+        web3.vap.getBalance(accountsResult[0], (balanceError, balanceResult) => {
           assert.equal(balanceError, null);
           assert.equal(typeof balanceResult, 'object');
           assert.equal(balanceResult.toNumber(10) > 0, true);
@@ -220,11 +220,11 @@ describe('IpcProvider', () => {
     });
 
     it('should get coinbase and balance', (done) => {
-      web3.eth.getCoinbase((accountsError, accountResult) => {
+      web3.vap.getCoinbase((accountsError, accountResult) => {
         assert.equal(accountsError, null);
         assert.equal(typeof accountResult, 'string');
 
-        web3.eth.getBalance(accountResult, (balanceError, balanceResult) => {
+        web3.vap.getBalance(accountResult, (balanceError, balanceResult) => {
           assert.equal(balanceError, null);
           assert.equal(typeof balanceResult, 'object');
           assert.equal(balanceResult.toNumber(10) > 0, true);