+++ CHANGELOG.md
@@ -1,3 +1,3 @@
-# 0.0.1 -- ethjs-provider-ipc
-1. Copied and adapted from ethjs-provider-http
+# 0.0.1 -- vapjs-provider-ipc
+1. Copied and adapted from vapjs-provider-http
 2. Added a proxy to test with TestRPC through a unix socket